requirements.txt
# --- For-Bot-Working --------- #
pyrogram
TgCrypto
pyromod==1.5
motor==2.5.1
python-dotenv
# --- For-Database ------------ #
pymongo
dnspython
# --- For-Web-Response ------- #
aiohttp
